Move attach-debug classes to main module
